# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PVT OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/role": {
            "get": {
                "tags": [
                    "ROLES"
                ],
                "summary": "LISTADO DE ROLES",
                "operationId": "getRoles",
                "parameters": [
                    {
                        "name": "display_name",
                        "in": "query",
                        "description": "Nombre del Rol",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/role/{role}": {
            "get": {
                "tags": [
                    "ROLES"
                ],
                "summary": "DETALLE DEL ROL",
                "operationId": "getROle",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/role/{role}/permission": {
            "patch": {
                "tags": [
                    "ROLES"
                ],
                "summary": "ESTABLECER O ELIMINAR EL PERMISO A UN ROL",
                "operationId": "getPermissionsByRole",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "permission_id": {
                                        "description": "nombres required",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/role/{role}/role_permissions": {
            "get": {
                "tags": [
                    "ROLES"
                ],
                "summary": "LISTADO DE PERMISOS ASIGNADOS A UN ROL",
                "operationId": "role_permissions",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id permiso",
                        "required": false
                    },
                    {
                        "name": "operation_id",
                        "in": "query",
                        "description": "id de la operacion",
                        "required": false
                    },
                    {
                        "name": "action_id",
                        "in": "query",
                        "description": "id de la acción ",
                        "required": false
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtro por nombre de permiso ",
                        "required": false
                    },
                    {
                        "name": "display_name",
                        "in": "query",
                        "description": "Filtro del nombre de visualización del permiso",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina a mostrar",
                        "required": false,
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Por Pagina",
                        "required": false,
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/global/city": {
            "get": {
                "tags": [
                    "CIUDAD"
                ],
                "summary": "LISTADO DE CIUDADES",
                "operationId": "getCiudades",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}